Rename dir `tools` to `utils`
