+++ audiobridge/bot/playlistHandler.py
@@ -4,10 +4,10 @@
 import logging
 import yt_dlp
 
-from audiobridge.tools.customErrors import CustomError, specific_errors
-from audiobridge.tools.sayOrReply import sayOrReply
-from audiobridge.tools.deleteMsg import deleteMsg
-from audiobridge.tools.yt_dlpShell import Yt_dlpShell
+from audiobridge.utils.customErrors import CustomError, specific_errors
+from audiobridge.utils.sayOrReply import sayOrReply
+from audiobridge.utils.deleteMsg import deleteMsg
+from audiobridge.utils.yt_dlpShell import Yt_dlpShell
 
 from audiobridge.config.bot import cfg as bot_cfg
 from audiobridge.config.handler import vars, WorkerTask
